サインイン、リセットページのコンポーネント分割および調整完了
