+++ src/components/organisms/SignInForm.tsx
@@ -1,22 +1,22 @@
 /** 外部import */
-import { ChangeEvent, FC, useCallback, useState, useContext } from 'react';
-import styled from 'styled-components';
 import { signInWithEmailAndPassword } from 'firebase/auth';
 import { doc, getDoc } from 'firebase/firestore';
-import { Link, useNavigate } from 'react-router-dom';
+import { ChangeEvent, FC, useCallback, useContext, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import styled from 'styled-components';
 
 /** 内部import */
-import type { User } from '../../types/Context/user';
-import { auth, db } from '../../firebase/index';
-import { AuthButton } from '../molecules';
-import { colorVariables as c } from '../../style';
+import { auth, db } from '../../firebase';
 import { isValidEmailFormat, isValidRequiredInput } from '../../lib';
 import { UserContext } from '../../providers/UserProvider';
+import { User } from '../../types/Context/user';
+import { H2Title, Input, TextLink } from '../atoms';
+import { AuthButton } from '../molecules';
 
 /** types */
 type SignInParams = (email: string, password: string) => void;
 
-export const SignInScreen: FC = () => {
+export const SignInForm: FC = () => {
   const navigate = useNavigate();
   const { setUser } = useContext(UserContext);
 
@@ -111,36 +111,35 @@ export const SignInScreen: FC = () => {
   };
 
   return (
-    <SBody>
-      <SInfoWrapper>
-        <SLogo>JACKETFLIX</SLogo>
-        <STitleWrapper>
-          <STitle>サインイン</STitle>
-        </STitleWrapper>
+    <>
+      <STitleWrapper>
+        <H2Title fontSize={'clamp(24px, 4.5vw, 32px)'}>サインイン</H2Title>
+      </STitleWrapper>
 
-        <SFormWrapper>
-          {/* メールアドレス */}
-          <SInputField>
-            <SLabelWrapper>
-              <SLabel htmlFor={'email'}>メールアドレス</SLabel>
-            </SLabelWrapper>
-            <SInputWrapper>
-              <SInput type={'email'} id={'email'} value={email} onChange={inputEmail} />
-            </SInputWrapper>
-          </SInputField>
+      <SFormWrapper>
+        {/* メールアドレス */}
+        <SInputField>
+          <Input
+            label={'メールアドレス'}
+            htmlFor={'email'}
+            value={email}
+            onChange={inputEmail}
+            type={'email'}
+            id={'email'}
+          />
+        </SInputField>
 
-          {/* パスワード */}
-          <SInputField>
-            <SLabelWrapper>
-              <SLabel htmlFor={'password'}>
-                パスワード <span>※6文字以上</span>
-              </SLabel>
-            </SLabelWrapper>
-            <SInputWrapper>
-              <SInput type={'password'} id={'password'} value={password} onChange={inputPassword} />
-            </SInputWrapper>
-          </SInputField>
-        </SFormWrapper>
+        {/* パスワード */}
+        <SInputField>
+          <Input
+            label={['パスワード', <span key={'spanKey'}>※6文字以上</span>]}
+            htmlFor={'password'}
+            value={password}
+            onChange={inputPassword}
+            type={'password'}
+            id={'password'}
+          />
+        </SInputField>
 
         {/* サインインボタン */}
         <AuthButton
@@ -152,41 +151,17 @@ export const SignInScreen: FC = () => {
 
         {/* 各種ページリンク */}
         <SLinkWrapper>
-          <SLink to={'/signup'}>アカウントをお持ちでない方はこちら</SLink>
+          <TextLink link={'/signup'}>アカウントをお持ちでない方はこちら</TextLink>
         </SLinkWrapper>
         <SLinkWrapper>
-          <SLink to={'/signin/reset'}>パスワードを忘れた方はこちら</SLink>
+          <TextLink link={'/signin/reset'}>パスワードを忘れた方はこちら</TextLink>
         </SLinkWrapper>
-      </SInfoWrapper>
-    </SBody>
+      </SFormWrapper>
+    </>
   );
 };
 
 /** style */
-const SBody = styled.div`
-  width: 100%;
-  min-height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding-top: 40px;
-  padding-bottom: 40px;
-`;
-
-const SInfoWrapper = styled.div`
-  padding-left: 16px;
-  padding-right: 16px;
-  width: 100%;
-  max-width: 600px;
-  text-align: center;
-`;
-
-const SLogo = styled.h1`
-  font-size: clamp(40px, 7.5vw, 56px);
-  font-weight: bold;
-  color: ${c.point};
-`;
-
 const STitleWrapper = styled.div`
   margin-top: 16px;
 
@@ -197,11 +172,6 @@ const STitleWrapper = styled.div`
   @media (min-width: 1024px) {
     margin-top: 32px;
   }
-`;
-
-const STitle = styled.h2`
-  font-size: clamp(24px, 4.5vw, 32px);
-  font-weight: bold;
 `;
 
 const SFormWrapper = styled.div`
@@ -216,76 +186,6 @@ const SInputField = styled.div`
   }
 `;
 
-const SLabelWrapper = styled.dt`
-  width: 100%;
-  margin-top: 0;
-  text-align: left;
-`;
-
-const SLabel = styled.label`
-  font-weight: bold;
-  font-size: 16px;
-
-  span {
-    font-size: 12px;
-    color: ${c.point};
-  }
-
-  @media (min-width: 768px) {
-    font-size: 18px;
-  }
-`;
-
-const SInputWrapper = styled.dd`
-  width: 100%;
-  margin-top: 8px;
-`;
-
-const SInput = styled.input`
-  font-family: inherit;
-  font-weight: normal;
-  width: 100%;
-  height: 48px;
-  padding-left: 14px;
-  padding-right: 14px;
-  border-radius: 8px;
-  color: #000;
-  appearance: none;
-  background-color: #fff;
-  font-size: 16px;
-
-  @media (min-width: 768px) {
-    font-size: 18px;
-  }
-
-  &:hover {
-    outline: none;
-    border: 3px solid ${c.point};
-    border-radius: 8px;
-  }
-
-  &:focus {
-    outline: none;
-    border: 3px solid ${c.point};
-    border-radius: 8px;
-  }
-
-  &:focus-visible {
-    outline: none;
-    border: 3px solid ${c.point};
-    border-radius: 8px;
-  }
-`;
-
 const SLinkWrapper = styled.p`
   margin-top: 32px;
-`;
-
-const SLink = styled(Link)`
-  color: ${c.secondary};
-  text-decoration: underline;
-
-  &:hover {
-    text-decoration: none;
-  }
 `;
