+++ src/components/pages/Reset.tsx
@@ -1,19 +1,14 @@
 /** 外部import */
 import { FC } from 'react';
+import { ResetForm } from '../organisms';
 
 /** 内部import */
-import { ResetScreen } from '../templates';
+import { AuthScreen } from '../templates';
 
 export const Reset: FC = () => {
   return (
-    <ResetScreen
-      logo={'JACKETFLIX'}
-      pageTitle={'パスワードリセット'}
-      leadText={[
-        'メールアドレスを入力し、「リセットする」ボタンを押してください。',
-        <br key="leadTextBr" />,
-        '入力されたアドレス宛にパスワードリセット用のメールをお送りします。',
-      ]}
-    />
+    <AuthScreen logo={'JACKETFLIX'}>
+      <ResetForm />
+    </AuthScreen>
   );
 };
